+++ src/dms_provider_bridge/drivers/webdav.py
@@ -10,7 +10,7 @@ from urllib.parse import quote, unquote, urlparse
 from xml.etree import ElementTree
 
 from dms_provider_bridge.core.config_loader import load_driver_config
-from dms_provider_bridge.core.debug import provider_debug_logger
+from dms_provider_bridge.core.debug import connection_debug_logger
 from dms_provider_bridge.core.errors import ProviderOperationError
 from dms_provider_bridge.models.bridge import BridgeAuthContext
 from dms_provider_bridge.models.item import DmsItem
@@ -31,7 +31,7 @@ class WebdavProvider(TcVfsContract):
         self.name = name or self.name
         self.config = config or load_driver_config("webdav")
         self.upstream_auth_scheme = self._configured_auth_scheme()
-        self.debug_logger = provider_debug_logger(self.name, self.config)
+        self.debug_logger = connection_debug_logger(self.name, self.config)
 
     def _configured_auth_scheme(self) -> str:
         credentials = self.config.get("credentials")